Update examples to use with_methods to create ctor
Use napi_create_constructor_for_wrap_with_methods to create constructor
function objects that have methods hooked up on their .prototype object.

This should be the way to create a constructor function object with
methods for performance reasons on V8.  The previous method is slow and
thus misleading to have as the example code for how to write a native
addon with NAPI.